+++ Sources/KeyboardKit/Autocomplete/AutocompleteContext.swift
@@ -3,13 +3,13 @@
 //  KeyboardKit
 //
 //  Created by Daniel Saidi on 2020-09-12.
-//  Copyright © 2021 Daniel Saidi. All rights reserved.
+//  Copyright © 2020-2023 Daniel Saidi. All rights reserved.
 //
 
 import Combine
 
 /**
- This is an observable class that can be used to store a set
+ This is an observab le class that can be used to store a set
  of autocomplete suggestions.
  
  KeyboardKit automatically creates an instance of this class